# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: openGym API
  version: 1.3.10
  summary: Passkey (WebAuthn) auth + per-user state storage for openGym
  description: |
    The backend of [openGym](https://opengym.ch), a free, open-source
    gym & body-weight tracker. A plain Node server, no framework, JSON-file storage,
    HMAC-signed session cookies.

    ### Authentication
    Sign-in is by **passkey** (WebAuthn, via `@simplewebauthn/server`). A successful
    `register/verify` or `login/verify` sets the session cookie; every authenticated
    route then works with that cookie. The paired mobile app has no shared cookie jar,
    so it redeems a short pairing code (`/api/pair/redeem`) for the *same* signed token
    and sends it as `Authorization: Bearer <token>` instead.

    An instance started with `PASSWORD_LOGIN=1` also offers **name-and-password
    sign-in** (tag `password`): opt-in per profile, scrypt-hashed, throttled, with an
    admin-issued one-time reset code instead of e-mail. A profile may also add an e-mail
    address to sign in with instead of its name (`/api/account/email`); no mail is ever
    sent to it. Its routes set exactly the same cookie. With the flag off (the default)
    every one of them answers 404.

    ### Throttling
    The `password` routes are counted per client address: at most 60 requests a minute
    across them, and wrong answers (a password, a reset code, an invite code on password
...
    counted by /64. Counters live in memory and reset on restart.

...

...

...

...
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/food:
    get:
      tags: [food]
      operationId: listFood
      summary: List food log entries in a date range
      description: >-
        Entries of the caller's profile between `from` and `to` (inclusive, at most 366 days),
        oldest first. Paged with `limit` (1-200, default 100) and the opaque `cursor` from the
        previous page's `next_cursor`. Unknown query parameters are refused. Also reachable with
        service auth (`X-OpenGym-Service-Token` + `X-OpenGym-Profile`) from the homelab MCP service.
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        '200':
          description: One page of entries.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/FoodLog' } }
                  next_cursor: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/FoodInvalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [food]
      operationId: createFood
      summary: Log a food entry
      description: >-
        Creates one entry. Unknown fields are refused (400, `field` names the first one). Body
        limit 16 KiB. With `idempotency_key` a retried request returns the original entry (200)
        instead of a duplicate; the same key with a different body answers 409
        `idempotency-key-reuse`. `source` is set by the server (`ui` for a session, `mcp` for
        service auth). 120 writes a minute per profile.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FoodLogCreate' }
            example: { date: "2026-10-09", time: "08:15", meal: breakfast, name: "Oatmeal", quantity: 80, unit: g, calories: 300, protein_g: 10, carbs_g: 54, fat_g: 6 }
      responses:
        '201':
          description: Created.
          content: { application/json: { schema: { $ref: '#/components/schemas/FoodLog' } } }
        '200':
          description: An idempotent replay; the entry created by the first request.
          content: { application/json: { schema: { $ref: '#/components/schemas/FoodLog' } } }
        '400': { $ref: '#/components/responses/FoodInvalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: The idempotency key was already used for a different entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: "idempotency key already used for a different entry", code: "idempotency-key-reuse" }
        '413': { description: Body over 16 KiB. }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/food/summary:
    get:
      tags: [food]
      operationId: foodSummary
      summary: Nutrition totals per day and for a range
      description: >-
        Calories and macros summed per logged day and over the whole range (at most 366 days),
        with the profile's goals. `entries_missing_calories` counts entries logged without a
        calorie value, so a low total can be told apart from an incomplete one.
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date } }
        - { name: to, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200':
          description: The summary.
          content: { application/json: { schema: { $ref: '#/components/schemas/NutritionSummary' } } }
        '400': { $ref: '#/components/responses/FoodInvalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/food/goals:
    get:
      tags: [food]
      operationId: getFoodGoals
      summary: The profile's daily nutrition goals
      responses:
        '200':
          description: The goals, or null when none are set.
          content:
            application/json:
              schema:
                type: object
                properties:
                  goals: { oneOf: [{ $ref: '#/components/schemas/NutritionGoals' }, { type: 'null' }] }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [food]
      operationId: setFoodGoals
      summary: Set the daily nutrition goals
      description: >-
        Replaces all four goals; an omitted or null field clears that goal. Session only: service
        auth answers 403 here.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NutritionGoalsInput' }
            example: { calories: 2400, protein_g: 160, carbs_g: 260, fat_g: 70 }
      responses:
        '200':
          description: The stored goals.
          content:
            application/json:
              schema:
                type: object
                properties:
                  goals: { $ref: '#/components/schemas/NutritionGoals' }
        '400': { $ref: '#/components/responses/FoodInvalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/food/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [food]
      operationId: getFood
      summary: One food log entry
      responses:
        '200':
          description: The entry.
          content: { application/json: { schema: { $ref: '#/components/schemas/FoodLog' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: 'No such entry on this profile (also the answer for another profile''s entry).' }
    put:
      tags: [food]
      operationId: updateFood
      summary: Update a food log entry
      description: >-
        Partial update: only the fields sent change; `null` clears an optional field. `version`
        is required and must equal the entry's current version, otherwise 409 with the current
        entry. Unknown fields are refused.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FoodLogUpdate' }
            example: { version: 1, calories: 320 }
      responses:
        '200':
          description: The updated entry (version incremented).
          content: { application/json: { schema: { $ref: '#/components/schemas/FoodLog' } } }
        '400': { $ref: '#/components/responses/FoodInvalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: No such entry on this profile. }
        '409': { $ref: '#/components/responses/FoodConflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      tags: [food]
      operationId: deleteFood
      summary: Delete a food log entry
      description: With `?version=N` the delete only happens if the entry is still at that version.
      parameters:
        - { name: version, in: query, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Deleted.
          content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } }
        '400': { $ref: '#/components/responses/FoodInvalid' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: No such entry on this profile. }
        '409': { $ref: '#/components/responses/FoodConflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gymsid
      description: |
        HMAC-signed session cookie, set by `register/verify` / `login/verify`.
        On HTTPS deployments it is issued as **`__Host-gymsid`** (host-locked,
        Secure); the unprefixed `gymsid` is what a plain-http `localhost` instance
        uses, and is still accepted for pre-upgrade sessions. HttpOnly + SameSite=Lax;
        browser clients never handle the value directly.
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        The same signed session token, carried as `Authorization: Bearer <token>` by
        the paired mobile app (obtained from `POST /api/pair/redeem`). Bearer requests
        skip the CSRF origin check; a browser never attaches this header on its own.
    serviceToken:
      type: apiKey
      in: header
      name: X-OpenGym-Service-Token
      description: >-
        Shared secret for the isolated MCP service. Only explicit service routes accept it;
        writes outside food and MCP link redemption are refused by the API.
    serviceProfile:
      type: apiKey
      in: header
      name: X-OpenGym-Profile
      description: The profile id resolved from the OAuth token; required on profile-scoped service routes.
  responses:
    Unauthorized:
      description: No session, expired/revoked session, or a disabled account.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "not signed in" }
...
    FoodInvalid:
      description: A field or query parameter is unknown, missing or out of bounds; `field` names it.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              field: { type: string }
          example: { error: "unknown field", field: "brand" }
    FoodConflict:
      description: The entry changed since it was read; `current` is the stored entry.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              code: { type: string, enum: [version-conflict] }
              current: { $ref: '#/components/schemas/FoodLog' }
...
    TooManyRequests:
      description: >-
        Paused by the sign-in throttle (see *Throttling* in the description). The request
        was not looked at; try again after `Retry-After` seconds.
      headers:
        Retry-After:
          description: Seconds until the pause ends.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "too many attempts, try again later", code: "locked", retryAfter: 60 }
...
  schemas:
    FoodLogCreate:
      type: object
      additionalProperties: false
      required: [date, meal, name]
      properties:
        date: { type: string, format: date, description: 'Local calendar date, YYYY-MM-DD (2000-2100).' }
        time: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
        meal: { type: string, enum: [breakfast, lunch, dinner, snack] }
        name: { type: string, minLength: 1, maxLength: 120 }
        quantity: { type: number, exclusiveMinimum: 0, maximum: 10000 }
        unit: { type: string, enum: [g, ml, piece, serving, cup, tbsp, tsp, oz] }
        calories: { type: number, minimum: 0, maximum: 20000 }
        protein_g: { type: number, minimum: 0, maximum: 2000 }
        carbs_g: { type: number, minimum: 0, maximum: 2000 }
        fat_g: { type: number, minimum: 0, maximum: 2000 }
        notes: { type: string, maxLength: 500 }
        idempotency_key: { type: string, pattern: '^[A-Za-z0-9._:-]{1,100}$' }
    FoodLogUpdate:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, minimum: 1 }
        date: { type: string, format: date }
        time: { type: [string, 'null'] }
        meal: { type: string, enum: [breakfast, lunch, dinner, snack] }
        name: { type: string, minLength: 1, maxLength: 120 }
        quantity: { type: [number, 'null'] }
        unit: { type: [string, 'null'], enum: [g, ml, piece, serving, cup, tbsp, tsp, oz, null] }
        calories: { type: [number, 'null'] }
        protein_g: { type: [number, 'null'] }
        carbs_g: { type: [number, 'null'] }
        fat_g: { type: [number, 'null'] }
        notes: { type: [string, 'null'], maxLength: 500 }
    FoodLog:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        date: { type: string, format: date }
        time: { type: [string, 'null'] }
        meal: { type: string, enum: [breakfast, lunch, dinner, snack] }
        name: { type: string }
        quantity: { type: [number, 'null'] }
        unit: { type: [string, 'null'] }
        calories: { type: [number, 'null'] }
        protein_g: { type: [number, 'null'] }
        carbs_g: { type: [number, 'null'] }
        fat_g: { type: [number, 'null'] }
        notes: { type: [string, 'null'] }
        source: { type: string, enum: [ui, mcp] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer }
        idempotency_key: { type: [string, 'null'] }
    NutritionGoalsInput:
      type: object
      additionalProperties: false
      properties:
        calories: { type: [number, 'null'], exclusiveMinimum: 0, maximum: 20000 }
        protein_g: { type: [number, 'null'], exclusiveMinimum: 0, maximum: 2000 }
        carbs_g: { type: [number, 'null'], exclusiveMinimum: 0, maximum: 2000 }
        fat_g: { type: [number, 'null'], exclusiveMinimum: 0, maximum: 2000 }
    NutritionGoals:
      allOf:
        - $ref: '#/components/schemas/NutritionGoalsInput'
        - type: object
          properties:
            updated_at: { type: string, format: date-time }
    NutritionSummary:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        days_logged: { type: integer }
        totals:
          type: object
          properties:
            entries: { type: integer }
            calories: { type: number }
            protein_g: { type: number }
            carbs_g: { type: number }
            fat_g: { type: number }
        days:
          type: array
          items:
            type: object
            properties:
              date: { type: string, format: date }
              entries: { type: integer }
              calories: { type: number }
              protein_g: { type: number }
              carbs_g: { type: number }
              fat_g: { type: number }
              entries_missing_calories: { type: integer }
        goals: { oneOf: [{ $ref: '#/components/schemas/NutritionGoals' }, { type: 'null' }] }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable message. This is the only error shape the API has.
        code:
          type: string
          description: >-
            A stable reason on the password, throttle and media routes, for a client to word
            in its own language (`bad-credentials`, `locked`, `too-short`, `media-quota`, …).
            Media refusals add the numbers to word it with (`maxMB`, `maxSec`, `usedMB`,
            `quotaMB`).
        retryAfter:
          type: integer
          description: 429 only. The same number of seconds as the Retry-After header.
    Ok:
      type: object
      properties:
        ok: { type: boolean, const: true }
...